requirements.txt
fastapi
uvicorn
requests
openai
python-dotenv
groq
